+++ src/main/java/roomescape/theme/repository/ThemeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.repository;
+package roomescape.theme.repository;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -7,7 +7,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import roomescape.reservation.domain.Theme;
+import roomescape.theme.domain.Theme;
 
 public interface ThemeRepository extends JpaRepository<Theme, Long> {
