+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.service;
+package roomescape.theme.service;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -7,11 +7,11 @@ import org.springframework.stereotype.Service;
 
 import roomescape.common.exception.AlreadyInUseException;
 import roomescape.common.exception.EntityNotFoundException;
-import roomescape.reservation.domain.Theme;
-import roomescape.reservation.dto.request.ThemeRequest;
-import roomescape.reservation.dto.response.ThemeResponse;
+import roomescape.theme.domain.Theme;
+import roomescape.theme.dto.request.ThemeRequest;
+import roomescape.theme.dto.response.ThemeResponse;
 import roomescape.reservation.repository.ReservationRepository;
-import roomescape.reservation.repository.ThemeRepository;
+import roomescape.theme.repository.ThemeRepository;
 
 @Service
 public class ThemeService {
